requirements.txt
pandas
matplotlib
nltk
scikit-learn
notebook
torch
mlflow
datasets
sacrebleu
transformers
rouge_score
py7zr
tqdm
PyYAML
pytest
boto3
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
uvicorn==0.18.3
fastapi==0.78.0
Jinja2==3.1.2

